requirements.txt
Pillow>=8.0.0
requests>=2.23